+++ authentication_lib/src/main/java/org/octri/authentication/server/security/service/PasswordResetTokenService.java
@@ -98,20 +98,6 @@ public class PasswordResetTokenService {
 	}
 
 	/**
-	 * Generates a password reset token for a user. Tokens are persisted in the database.
-	 *
-	 * @deprecated The overload that takes a {@link Duration} should be used instead.
-	 * @param user
-	 * @param expireInMinutes
-	 * @return Returns a new {@link PasswordResetToken} for the given user.
-	 */
-	@Deprecated(forRemoval = true, since = "1.2.0")
-	public PasswordResetToken generatePasswordResetToken(final User user, Integer expireInMinutes) {
-		Assert.notNull(user, "User cannot be null");
-		return save(new PasswordResetToken(user, expireInMinutes));
-	}
-
-	/**
 	 * Generates a password reset token for a user that will be valid for the specified duration. Tokens are persisted
 	 * in the database.
 	 *